text-to-PPT-project/requirements.txt
openai 
python-pptx